requirements.txt
playwright
google-generativeai
openai
anthropic
python-dotenv
PyPDF2
jinja2
beautifulsoup4
requests
python-slugify
pydantic
pydantic-settings
tenacity
